requirements.txt
python-binance==1.0.19
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
requests==2.31.0
